ref(store): Move common code into own module
