+++ server/src/endpoints/mod.rs
@@ -3,6 +3,8 @@
 //! This module contains implementations for all supported relay endpoints, as well as a generic
 //! `forward` endpoint that sends unknown requests to the upstream.
 
+mod common;
+
 pub mod events;
 pub mod forward;
 pub mod healthcheck;
